requirements.txt
asgiref==3.11.0
charset-normalizer==3.4.4
ciscoconfparse==1.9.52
colorama==0.4.6
Deprecated==1.3.1
Django==6.0.1
dnspython==2.8.0
gunicorn==21.2.0
hier-config==2.2.3
loguru==0.7.2
passlib==1.7.4
pillow==12.1.0
PyYAML==6.0.3
reportlab==4.4.9
sqlparse==0.5.5
toml==0.10.2
tzdata==2025.3
win32_setctime==1.2.0
wrapt==2.0.1
